refactor: satisfy SwiftLint with splits and small fixes
- Extract ReviewCuratedThemeMap and WeeklyReviewAggregatesBuilder theme
  sections into dedicated files; trim WeeklyReviewAggregates.
- Move insight card loading UI to ReviewInsightThemeCardSupport+Loading.
- Use extensions on ReviewScreen, SequentialSectionPrimaryColumn, and
  several XCTest types to keep type bodies under thresholds.
- Fix identifier_name and line_length issues; narrow lint disables for
  concept engine and long UI test.

Made-with: Cursor
