+++ GraceNotes/GraceNotes/Features/Journal/Services/ReviewCuratedThemeMap.swift
@@ -1,0 +1,166 @@
+import Foundation
+
+struct ReviewCuratedThemeMap {
+    let globalAliases: [String: String]
+    let sectionOverrides: [ReviewThemeSourceCategory: [String: String]]
+    let hardBannedConcepts: Set<String>
+    let penalizedConcepts: Set<String>
+    let crossLanguageAliases: [String: String]
+    let canonicalDisplayLabels: [String: String]
+
+    static let defaultMap = ReviewCuratedThemeMap(
+        globalAliases: [
+            "sleep": "sleep",
+            "better sleep": "sleep",
+            "slept": "sleep",
+            "rest": "rest",
+            "recover": "rest",
+            "recovery": "rest",
+            "downtime": "rest",
+            "quiet": "quiet time",
+            "quiet time": "quiet time",
+            "stillness": "quiet time",
+            "calm": "calm",
+            "peace": "calm",
+            "focused": "focus",
+            "focus": "focus",
+            "concentration": "focus",
+            "boundaries": "boundaries",
+            "boundary": "boundaries",
+            "time boundaries": "boundaries",
+            "exercise": "movement",
+            "workout": "movement",
+            "training": "movement",
+            "walk": "walking",
+            "walking": "walking",
+            "morning walk": "walking",
+            "walks": "walking",
+            "run": "movement",
+            "running": "movement",
+            "stretching": "movement",
+            "scripture": "scripture",
+            "bible": "scripture",
+            "devotional": "scripture",
+            "family dinner": "family connection",
+            "family time": "family connection",
+            "time with family": "family connection",
+            "friends": "friendship",
+            "friend": "friendship",
+            "friendship": "friendship",
+            "partner": "partner connection",
+            "husband": "partner connection",
+            "wife": "partner connection",
+            "spouse": "partner connection",
+            "parenting": "parenting",
+            "kids": "parenting",
+            "children": "parenting",
+            "work stress": "work pressure",
+            "work pressure": "work pressure",
+            "busy at work": "work pressure",
+            "finances": "financial pressure",
+            "money stress": "financial pressure",
+            "budget": "financial pressure",
+            "home chores": "home care",
+            "house chores": "home care",
+            "cleaning": "home care",
+            "meal prep": "home care",
+            "therapy": "therapy",
+            "counseling": "therapy",
+            "nature": "nature",
+            "sunlight": "nature",
+            "outside": "nature",
+            "reading": "learning",
+            "study": "learning",
+            "learning": "learning",
+            "creative work": "creativity",
+            "writing": "creativity",
+            "drawing": "creativity"
+        ],
+        sectionOverrides: [
+            .needs: [
+                "time alone": "personal time",
+                "alone time": "personal time",
+                "quiet morning": "quiet time",
+                "less pressure": "work pressure",
+                "space to breathe": "calm",
+                "clear priorities": "focus"
+            ],
+            .gratitudes: [
+                "good weather": "nature",
+                "sunny weather": "nature",
+                "warm sunlight": "nature",
+                "my routine": "daily rhythm",
+                "morning routine": "daily rhythm"
+            ],
+            .people: [
+                "mom": "mom",
+                "mother": "mom",
+                "mama": "mom",
+                "dad": "dad",
+                "father": "dad",
+                "爸": "dad",
+                "妈妈": "mom",
+                "爸爸": "dad"
+            ]
+        ],
+        hardBannedConcepts: [
+            "today", "this week", "week", "thing", "things", "stuff", "something", "anything",
+            "nothing", "life", "moment", "moments", "reflection", "reflections", "entry",
+            "journal", "note", "notes", "feeling", "feelings",
+            "need", "needs", "needed",
+            "gratitude", "grateful", "thankful", "thankfulness", "thanks", "gratitude practice",
+            "pray", "prayer", "praying", "prayed",
+            "感恩", "祷告"
+        ],
+        penalizedConcepts: [
+            "work", "family", "health", "relationship", "home", "time", "people"
+        ],
+        crossLanguageAliases: [
+            "睡眠": "sleep",
+            "休息": "rest",
+            "安静": "quiet time",
+            "平静": "calm",
+            "专注": "focus",
+            "边界": "boundaries",
+            "散步": "walking",
+            "运动": "movement",
+            "圣经": "scripture",
+            "家人": "family connection",
+            "朋友": "friendship",
+            "伴侣": "partner connection",
+            "育儿": "parenting",
+            "工作压力": "work pressure",
+            "财务压力": "financial pressure",
+            "家务": "home care",
+            "大自然": "nature",
+            "学习": "learning",
+            "创造": "creativity"
+        ],
+        canonicalDisplayLabels: [
+            "sleep": "Sleep",
+            "rest": "Rest",
+            "quiet time": "Quiet time",
+            "calm": "Calm",
+            "focus": "Focus",
+            "boundaries": "Boundaries",
+            "movement": "Exercise",
+            "walking": "Walking",
+            "scripture": "Scripture",
+            "family connection": "Family connection",
+            "friendship": "Friendship",
+            "partner connection": "Partner connection",
+            "parenting": "Parenting",
+            "work pressure": "Work pressure",
+            "financial pressure": "Financial pressure",
+            "home care": "Home care",
+            "therapy": "Therapy",
+            "nature": "Nature",
+            "learning": "Learning",
+            "creativity": "Creativity",
+            "personal time": "Personal time",
+            "daily rhythm": "Daily rhythm",
+            "mom": "Mom",
+            "dad": "Dad"
+        ]
+    )
+}
